pyproject.toml
[project]
name = "quickshield"
version = "0.1.0"
description = "QuickShield CE: check website uptime, SSL expiry, DNS changes, and security headers."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Spencer Kemmet", email = "spencerkemmet@gmail.com" }]
dependencies = [
  "typer>=0.12.3",
  "PyYAML>=6.0.2",
]

[project.scripts]
quickshield = "quickshield.cli:app"
